+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 * Ability to open config in editor if config is missing
 * New config section for setting default target language
+* New *View in Google Translate* action on translations
 
 ### Changed
 * Subtext now displays language names instead of codes